pyproject.toml
[project]
name = "mcp-router-aliyun-test"
version = "0.1.0"
description = "MCP Router - A service discovery and proxy for MCP services"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "dashscope>=1.19.0",
    "dashvector>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
mcp-router = "mcp_router:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["mcp_router*"]

[tool.pytest.ini_options]
testpaths = ["."]
python_files = ["test_*.py"]
python_functions = ["test_*"]
